requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
playwright
requests
pypdf
